Remove use of dep and update version of Go to Go 1.13 (#8)
Remove use of `dep` and update version of Go to Go 1.13.

The `stellar/go` repo was changed in stellar/go#1542 to use the build-in Go Modules instead of Dep for dependency management. The standard `go` commands can now be relied on for downloading and installing dependencies without any need to run any additional commands like we had to with Dep. Go 1.13 is the latest version of Go and we should use it for these tests since the repository no longer supports Go 1.10.